# config/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "catalog-update",
        "version": "1.0",
        "description": "カタログ更新サービス"
    },
    "servers": [
        {
            "url": "/catalog-update"
        }
    ],
    "paths": {
...
        "/join": {
            "post": {
                "tags": [
                    "申請"
                ],
                "description": "Region参加要求を登録します。",
                "requestBody": {
                    "description": "参加要求",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "description": "APPまたはWFが参加要求を登録するリクエストボディ",
                                "properties": {
                                    "region": {
                                        "type": "object",
                                        "description": "region",
                                        "nullable": true,
                                        "properties": {
                                            "code": {
                                                "type": "integer",
                                                "description": "regionコード",
                                                "example": 1000003
                                            },
                                            "version": {
                                                "type": "integer",
                                                "description": "regionバージョン",
                                                "example": 1
                                            }
                                        }
                                    },
                                    "actor": {
                                        "type": "object",
                                        "description": "参加者",
                                        "nullable": true,
                                        "properties": {
                                            "code": {
                                                "type": "integer",
                                                "description": "actorコード",
                                                "example": 1000002
                                            },
                                            "version": {
                                                "type": "integer",
                                                "description": "actorバージョン",
                                                "example": 1
                                            }
                                        }
                                    },
                                    "isDraft": {
                                        "type": "boolean",
                                        "description": "下書きフラグ",
                                        "example": false
                                    }
                                },
                                "example": {
                                    "region": {
                                        "code": 1000011,
                                        "version": 1
                                    },
                                    "actor": {
                                        "code": 1000021,
                                        "version": 1,
                                        "app": null,
                                        "wf": [
                                            {
                                                "code": 1000021,
                                                "version": 1
                                            }
                                        ]
                                    },
                                    "isDraft": false
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "参加要求に成功した際のレスポンスコード",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "description": "参加要求のレスポンスボディ",
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "description": "id",
                                            "example": 1
                                        },
                                        "region": {
                                            "type": "object",
                                            "description": "region",
                                            "properties": {
                                                "code": {
                                                    "type": "integer",
                                                    "description": "regionコード",
                                                    "example": 1000003
                                                },
                                                "version": {
                                                    "type": "integer",
                                                    "description": "regionバージョン",
                                                    "example": 1
                                                }
                                            }
                                        },
                                        "actor": {
                                            "type": "object",
                                            "description": "参加者",
                                            "properties": {
                                                "code": {
                                                    "type": "integer",
                                                    "description": "actorコード",
                                                    "example": 1000002
                                                },
                                                "version": {
                                                    "type": "integer",
                                                    "description": "actorバージョン",
                                                    "example": 1
                                                }
                                            }
                                        },
                                        "expireAt": {
                                            "type": "string",
                                            "description": "有効期限",
                                            "example": "2020-01-10T10:00:00.999+0900"
                                        },
                                        "isDraft": {
                                            "type": "boolean",
                                            "description": "下書きフラグ",
                                            "example": true
                                        }
                                    },
                                    "example": {
                                        "id": 1,
                                        "region": {
                                            "code": 1000011,
                                            "version": 1
                                        },
                                        "actor": {
                                            "code": 1000021,
                                            "version": 1,
                                            "app": null,
                                            "wf": [
                                                {
                                                    "code": 1000021,
                                                    "version": 1
                                                }
                                            ]
                                        },
                                        "expireAt": "2020-01-10 10:00:00.000",
                                        "isDraft": false
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "申請取得"
                ],
                "description": "Region参加申請を取得します。",
                "parameters": [
                    {
                        "name": "is_request",
                        "description": "自分が申請元かどうか",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        },
                        "required": false,
                        "example": false
                    },
                    {
                        "name": "in_approved",
                        "description": "承認済みを含む（Default:false）",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        },
                        "required": false,
                        "example": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Region参加申請取得に成功した際のレスポンスコード",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "description": "Region参加申請取得のレスポンスボディ",
                                    "items": {
                                        "$ref": "#/components/schemas/joinResponse"
                                    },
                                    "example": [
                                        {
                                            "id": 1,
                                            "type": 1,
                                            "caller": {
                                                "code": 1000021,
                                                "version": 1
                                            },
                                            "region": {
                                                "code": 1000011,
                                                "version": 1
                                            },
                                            "actor": {
                                                "code": 1000021,
                                                "version": 1
                                            },
                                            "expireAt": "2020-04-30T00:00:00.000+9:00",
                                            "isDraft": false,
                                            "status": 0,
                                            "comment": null,
                                            "approver": null,
                                            "approvalAt": null
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/join/remove": {
            "post": {
                "tags": [
                    "解除申請"
                ],
                "description": "Region離脱要求を登録します。",
                "requestBody": {
                    "description": "離脱要求",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "description": "APPまたはWFが離脱要求を登録するリクエストボディ",
                                "properties": {
                                    "region": {
                                        "type": "object",
                                        "description": "region",
                                        "nullable": true,
                                        "properties": {
                                            "code": {
                                                "type": "integer",
                                                "description": "regionコード",
                                                "example": 1000003
                                            },
                                            "version": {
                                                "type": "integer",
                                                "description": "regionバージョン",
                                                "example": 1
                                            }
                                        }
                                    },
                                    "actor": {
                                        "type": "object",
                                        "description": "離脱者",
                                        "nullable": true,
                                        "properties": {
                                            "code": {
                                                "type": "integer",
                                                "description": "actorコード",
                                                "example": 1000002
                                            },
                                            "version": {
                                                "type": "integer",
                                                "description": "actorバージョン",
                                                "example": 1
                                            }
                                        }
                                    },
                                    "isDraft": {
                                        "type": "boolean",
                                        "description": "下書きフラグ",
                                        "example": false
                                    }
                                },
                                "example": {
                                    "region": {
                                        "code": 1000011,
                                        "version": 1
                                    },
                                    "actor": {
                                        "code": 1000021,
                                        "version": 1,
                                        "app": null,
                                        "wf": [
                                            {
                                                "code": 1000021,
                                                "version": 1
                                            }
                                        ]
                                    },
                                    "isDraft": false
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "離脱要求に成功した際のレスポンスコード",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "description": "離脱要求のレスポンスボディ",
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "description": "id",
                                            "example": 1
                                        },
                                        "region": {
                                            "type": "object",
                                            "description": "region",
                                            "properties": {
                                                "code": {
                                                    "type": "integer",
                                                    "description": "regionコード",
                                                    "example": 1000003
                                                },
                                                "version": {
                                                    "type": "integer",
                                                    "description": "regionバージョン",
                                                    "example": 1
                                                }
                                            }
                                        },
                                        "actor": {
                                            "type": "object",
                                            "description": "離脱者",
                                            "properties": {
                                                "code": {
                                                    "type": "integer",
                                                    "description": "actorコード",
                                                    "example": 1000002
                                                },
                                                "version": {
                                                    "type": "integer",
                                                    "description": "actorバージョン",
                                                    "example": 1
                                                }
                                            }
                                        },
                                        "expireAt": {
                                            "type": "string",
                                            "description": "有効期限",
                                            "example": "2020-01-10T10:00:00.999+0900"
                                        },
                                        "isDraft": {
                                            "type": "boolean",
                                            "description": "下書きフラグ",
                                            "example": false
                                        }
                                    },
                                    "example": {
                                        "id": 1,
                                        "region": {
                                            "code": 1000011,
                                            "version": 1
                                        },
                                        "actor": {
                                            "code": 1000021,
                                            "version": 1,
                                            "app": null,
                                            "wf": [
                                                {
                                                    "code": 1000021,
                                                    "version": 1
                                                }
                                            ]
                                        },
                                        "expireAt": "2020-01-10 10:00:00.000",
                                        "isDraft": false
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/join/approval/{code}": {
            "post": {
                "tags": [
                    "承認結果登録"
                ],
                "description": "承認結果登録を登録します。",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "description": "承認申請コード",
                        "example": "123456"
                    }
                ],
                "requestBody": {
                    "description": "承認結果登録",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "description": "承認する場合のリクエストボディ",
                                "properties": {
                                    "status": {
                                        "type": "integer",
                                        "description": "承認結果",
                                        "example": 1
                                    },
                                    "comment": {
                                        "type": "string",
                                        "description": "コメント",
                                        "example": null
                                    }
                                },
                                "example": {
                                    "status": 1,
                                    "comment": null
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "承認結果登録のレスポンスコード",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "description": "承認結果登録レスポンスボディ",
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "description": "承認結果",
                                            "example": 1
                                        },
                                        "comment": {
                                            "type": "string",
                                            "description": "コメント",
                                            "example": null
                                        },
                                        "approvalActor": {
                                            "type": "object",
                                            "description": "approvalActor",
                                            "properties": {
                                                "_value": {
                                                    "type": "integer",
                                                    "description": "value",
                                                    "example": 1000004
                                                },
                                                "_ver": {
                                                    "type": "integer",
                                                    "description": "バージョン",
                                                    "example": 1
                                                }
                                            }
                                        },
                                        "approver": {
                                            "type": "string",
                                            "description": "承認者",
                                            "example": "region_root_member01"
                                        },
                                        "approvalAt": {
                                            "type": "string",
                                            "description": "承認日時",
                                            "example": "2020-01-10T10:00:00.999+0900"
                                        }
                                    },
                                    "example": {
                                        "status": 1,
                                        "comment": null,
                                        "approvalActor": {
                                            "_value": 1000004,
                                            "_ver": 1
                                        },
                                        "approver": "region_root_member01",
                                        "approvalAt": "2020-01-10T10:00:00.999+0900"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "joinResponse": {
                "type": "object",
                "description": "Region参加申請取得のレスポンスボディ",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "参加管理ID"
                    },
                    "type": {
                        "type": "integer",
                        "description": "申請種別"
                    },
                    "caller": {
                        "type": "object",
                        "description": "申請元アクター",
                        "properties": {
                            "code": {
                                "type": "integer",
                                "description": "申請元アクターカタログコード"
                            },
                            "version": {
                                "type": "integer",
                                "description": "申請元アクターカタログバージョン"
                            }
                        }
                    },
                    "region": {
                        "type": "object",
                        "description": "参加Region",
                        "properties": {
                            "code": {
                                "type": "integer",
                                "description": "参加Regionカタログコード"
                            },
                            "version": {
                                "type": "integer",
                                "description": "参加Regionカタログバージョン"
                            }
                        }
                    },
                    "actor": {
                        "type": "object",
                        "description": "参加アクター",
                        "properties": {
                            "code": {
                                "type": "integer",
                                "description": "参加アクターカタログコード"
                            },
                            "version": {
                                "type": "integer",
                                "description": "参加アクターカタログバージョン"
                            }
                        }
                    },
                    "expireAt": {
                        "type": "string",
                        "description": "承認有効期限"
                    },
                    "isDraft": {
                        "type": "boolean",
                        "description": "下書きフラグ"
                    },
                    "status": {
                        "type": "integer",
                        "description": "認定ステータス"
                    },
                    "comment": {
                        "type": "string",
                        "description": "コメント"
                    },
                    "approver": {
                        "type": "string",
                        "description": "承認者"
                    },
                    "approvalAt": {
                        "type": "string",
                        "description": "承認日時"
                    }
                }
            }
        }
    }
}